feat(API): Moving properties to relevant classes
